hw4/requirements.txt
astropy
numpy
pandas
scipy
Pillow
photutils
